added controller and interactor, moved around some result functions in searcher and implemented a UserInteractor to keep track of who is logged in, moved loginUser and forgot password to user controller
