+++ src/main/java/entities/GuestUser.java
@@ -2,7 +2,7 @@ package entities;
 
 import java.util.List;
 
-public class GuestUser extends User {
+public class GuestUser extends entities.User {
     public GuestUser(){
     }
 
@@ -10,7 +10,7 @@ public class GuestUser extends User {
         return true;
     }
 
-    public List<User> getUser(){
+    public List<entities.User> getUser(){
         return null;
     }
 
@@ -24,25 +24,5 @@ public class GuestUser extends User {
     public void deleteUser(){
     }
 
-    public void logInUser() throws Exception {
-        UserData u = new UserData();
-        // ask for email and password
-        Scanner email = new Scanner(System.in);
-        System.out.println("Enter your email: ");
-        Scanner password = new Scanner(System.in);
-        System.out.println("Enter your password: ");
-        // call on u.login
-        u.logInUser(String.valueOf(email), String.valueOf(password));
-    }
-    public void logoutUser(){
-    }
 
-    public void forgotPassword() {
-        // ask for email
-        Scanner emailInput = new Scanner(System.in);
-        System.out.println("Please enter your email address.");
-        // sends an email
-        System.out.println("An email to reset your password has been sent to " + emailInput.next());
-        // email leads to her instance of Entities.RegisteredUser
-    }
 }
