+++ src/main/java/interface_adapters/Searcher.java
@@ -1,5 +1,5 @@
 package interface_adapters;
-import use_cases.*;
+import use_cases.MusicData;
 
 import java.io.File;
 import java.io.FileNotFoundException;
@@ -48,31 +48,6 @@ public class Searcher {
         return getKeyListSortedByValue(scoreMap);
     }
 
-    /**
-     * Prints out relevant information based on the given action.
-     * Available actions: top, recommend
-     *
-     * @param action the action that User chose to take
-     */
-    public void actionResult(String action) {
-        if (action.startsWith("top")) {
-            String[] split = action.split(" ");
-            int amt = Integer.parseInt(split[1]);
-            List<String> artists = ArtistComparer.topArtistNames(amt);
-            for (int i = 0;i < artists.size();i++) {
-                System.out.println((i + 1) + ". " + artists.get(i));
-            }
-        }
-        else if (action.startsWith("recommend")) {
-            String genre = action.substring(10, action.indexOf("artist") - 1);
-            MusicData mD = new MusicData();
-            GuestUser gU = new GuestUser();
-            Artist artist = mD.recommendArtist(genre, false, gU);
-            for (Map.Entry<String, Object> entry : artist.getInfo().entrySet()) {
-                System.out.println(entry.getKey() + ": " + entry.getValue());
-            }
-        }
-    }
 
     /**
      * Returns an ArrayList of at most 10 most relevant Artist that the User can navigate based on the given keyword.
@@ -111,15 +86,7 @@ public class Searcher {
 
     }
 
-    /**
-     * Returns an ArrayList of Artists within the given genre.
-     *
-     * @param genre the name of the genre
-     * @return an ArrayList of Artist with the given genre
-     */
-    public List<Artist> genreResult(String genre) {
-        return MusicData.getArtistsByGenre(genre);
-    }
+
 
     /**
      * Returns relevant score which indicates how relevant the keyword is to the str.
@@ -162,6 +129,7 @@ public class Searcher {
         return sortedList;
     }
 
+
     public static void main(String[] args) throws FileNotFoundException {
         Searcher searcher = new Searcher();
         Scanner scanner = new Scanner(System.in);
@@ -182,7 +150,7 @@ public class Searcher {
                 try {
                     int index = Integer.parseInt(scanner.nextLine().strip());
                     if (1 <= index && index <= actions.size()) {
-                        searcher.actionResult(actions.get(index - 1));
+                        MusicData.actionResult(actions.get(index - 1));
                     }
                     else {
                         System.out.println("Invalid index");
@@ -207,7 +175,7 @@ public class Searcher {
                 try {
                     int index = Integer.parseInt(scanner.nextLine().strip());
                     if (1 <= index && index <= artists.size()) {
-                        Artist artist = MusicData.artistResult(artists.get(index - 1));
+                        Artist artist = artistResult(artists.get(index - 1));
                         for (Map.Entry<String, Object> entry : artist.getInfo().entrySet()) {
                             System.out.println(entry.getKey() + ": " + entry.getValue());
                         }
@@ -250,5 +218,6 @@ public class Searcher {
             System.out.println("Invalid search type, please enter keyword, genres or artist.");
         }
     }
+
 }
