+++ src/main/java/interface_adapters/UserController.java
@@ -1,0 +1,33 @@
+package interface_adapters;
+import use_cases.*;
+
+public class UserController {
+
+    public void setUser(String type) {
+
+    }
+
+    public void logInUser() throws Exception {
+        UserData u = new UserData();
+        // ask for email and password
+        Scanner email = new Scanner(System.in);
+        System.out.println("Enter your email: ");
+        Scanner password = new Scanner(System.in);
+        System.out.println("Enter your password: ");
+        // call on u.login
+        u.logInUser(String.valueOf(email), String.valueOf(password));
+    }
+    public void logoutUser(){
+    }
+
+    public void forgotPassword() {
+        // ask for email
+        Scanner emailInput = new Scanner(System.in);
+        System.out.println("Please enter your email address.");
+        // sends an email
+        System.out.println("An email to reset your password has been sent to " + emailInput.next());
+        // email leads to her instance of Entities.RegisteredUser
+    }
+
+
+}
