+++ src/main/java/use_cases/UserInteractor.java
@@ -1,0 +1,13 @@
+package use_cases;
+import entities.*;
+
+import java.util.HashMap;
+import java.util.List;
+import java.util.Map;
+
+
+public class UserInteractor {
+    public static User loggedinUser = new GuestUser();
+    public static boolean isGuestUser = true;
+
+}
