+++ src/main/java/use_cases/music_database/Data_2
@@ -1,2 +1,2 @@
-entities.Artist, Streams, Follows, Genre, Likes
+4. entities.Artist, Streams, Follows, Genre, Likes
 Drake, 512439845, 68195320, Hip-Hop/Rap, 1 1 0 1 1 0 0 0 0 1 0 1 0 1 0 1 1 0 0 1 0 0 1 1 1 1 1 0 0 1 1 1 1 1 1 0 1 0 0 0 1 0 1 0 1 1 1 1 0 0, Bad Bunny, 468019223, 58186026, Latin trap, 1 0 0 1 1 0 0 1 1 1 1 1 0 1 1 1 0 1 0 0 1 1 0 1 0 0 0 1 1 1 0 0 0 1 1 1 1 1 0 1 1 1 1 0 0 0 1 0 0 0, Ed Sheeran, 389301368, 104277661, Pop, 0 1 1 0 0 0 1 0 0 0 1 1 0 1 1 1 1 0 1 0 0 0 0 1 1 0 1 1 0 0 1 1 0 1 1 0 0 0 0 0 1 1 0 0 1 0 1 0 0 1, The Weeknd, 219547189, 40649338, R&B, 1 0 0 1 1 1 1 0 1 0 1 0 1 0 0 1 1 1 0 0 1 1 1 1 1 0 1 0 1 1 0 1 0 0 1 1 1 1 1 1 1 1 0 1 0 0 0 0 1 0, Taylor Swift, 353018224, 61126540, Pop, 0 0 0 0 0 1 1 0 0 1 0 0 0 0 1 0 1 0 0 0 0 1 1 1 0 1 0 0 1 1 1 1 1 1 0 0 1 0 0 1 0 0 0 0 0 0 0 1 1 0, Justin Bieber, 331517983, 66952961, Pop, 0 0 1 1 0 0 1 1 0 1 0 1 0 1 0 0 0 0 1 0 1 0 0 0 1 1 0 0 1 0 1 0 0 1 0 1 0 0 0 1 0 1 0 0 0 0 0 0 1 1, Ariana Grande, 328647066, 84367109, Pop, 0 1 0 1 1 0 1 0 1 1 1 1 1 0 1 1 0 0 1 1 0 1 0 0 1 1 1 1 0 0 1 1 1 0 0 1 0 1 1 1 0 1 1 1 1 0 0 1 0 1, Eminem, 311615147, 60930468, Hip-Hop/Rap, 0 1 0 0 1 1 0 1 1 0 1 0 0 1 1 1 0 0 1 1 1 1 0 1 1 1 0 0 0 0 1 0 1 0 1 1 1 1 1 1 1 0 0 1 0 0 1 1 1 0, Post Malone, 304014789, 38572656, Hip-Hop/Rap, 1 0 0 1 0 1 0 1 1 0 1 1 1 0 0 0 1 0 0 0 1 0 0 0 0 1 0 0 1 1 0 0 0 0 1 1 1 0 0 1 0 1 0 1 0 1 0 0 1 1, BTS, 294542407, 56599167, K-Pop, 0 0 1 1 0 1 1 0 1 0 0 0 0 1 0 0 1 0 1 0 1 0 1 0 0 1 1 1 1 1 1 1 1 0 0 1 0 0 1 1 0 1 0 1 0 0 1 0 0 0, J Balvin, 281496420, 35642927, Latin urbano, 0 0 0 0 1 0 0 1 0 0 1 0 1 1 0 1 0 0 1 0 1 1 0 0 0 1 1 0 0 1 1 0 0 0 0 0 1 1 0 1 1 0 0 1 1 0 0 0 0 1, Billie Eilish, 243232816, 70931638, Pop, 1 0 0 0 0 0 0 0 1 1 1 0 0 1 1 0 1 0 1 0 0 0 1 1 0 0 1 0 1 1 0 1 1 0 0 0 1 0 0 0 1 1 1 0 0 0 1 1 0 0, Khalid, 213702984, 15586617, R&B, 0 1 0 0 1 0 1 0 1 0 1 0 1 1 1 1 1 1 0 0 1 0 1 0 1 1 0 0 1 1 0 0 0 0 0 1 1 0 1 0 0 0 0 0 0 1 1 1 1 1, Rihanna, 212628727, 50706876, Pop, 0 1 0 1 0 0 1 1 1 0 1 0 0 1 0 1 1 1 1 1 0 0 0 0 0 0 0 0 0 0 1 1 1 1 0 1 0 0 0 0 1 1 1 0 1 0 0 1 0 0
